Refactor icons & styled tag/stack list components
- Moves icons used Resume to own directory for consumption by anyone
- Extracts markup & styles for lists of items w/ rounded borders & no
list style into own component
- Updates resume pages to reference newly extracted components
