+++ src/components/icons/WebAddressIcon.module.sass
