+++ src/components/resume/Shared.module.sass
@@ -16,9 +16,3 @@
 .infoList > li > svg
   height: 2rem
   width: 2rem
-
-svg.stroke
-  fill: transparent
-  stroke: colors.$dark
-  stroke-linecap: round
-  stroke-linejoin: miter
